+++ lib/widgets/button_quarter.dart
@@ -19,65 +19,35 @@ class ButtonQuarter extends StatefulWidget {
 class _ButtonQuarterState extends State<ButtonQuarter> {
   @override
   Widget build(BuildContext context) {
-    return Column(
-      children: [
-        const Text(
-          'Quarter',
-          style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-        ),
-        TextButton(
-            onPressed: () {
-              if (widget.sport == 1) {
-                context.read<QuarterBasketball>().update();
-              }
-              if (widget.sport == 2) {
-                context.read<QuarterVolleyball>().update();
-              }
-              if (widget.sport == 3) {
-                context.read<QuarterSoccer>().update();
-              }
-              if (widget.sport == 4) {
-                context.read<QuarterFutsal>().update();
-              }
-              if (widget.sport == 5) {
-                context.read<QuarterBadminton>().update();
-              }
-              if (widget.sport == 6) {
-                context.read<QuarterTabletennis>().update();
-              }
-            },
-            child: widget.sport == 1
-                ? Text(
-                    '${context.watch<QuarterBasketball>().getQuarter}',
-                    style: const TextStyle(fontSize: 30),
-                  )
-                : widget.sport == 2
-                    ? Text(
-                        '${context.watch<QuarterVolleyball>().getQuarter}',
-                        style: const TextStyle(fontSize: 30),
-                      )
-                    : widget.sport == 3
-                        ? Text(
-                            '${context.watch<QuarterSoccer>().getQuarter}',
-                            style: const TextStyle(fontSize: 30),
-                          )
-                        : widget.sport == 4
-                            ? Text(
-                                '${context.watch<QuarterFutsal>().getQuarter}',
-                                style: const TextStyle(fontSize: 30),
-                              )
-                            : widget.sport == 5
-                                ? Text(
-                                    '${context.watch<QuarterBadminton>().getQuarter}',
-                                    style: const TextStyle(fontSize: 30),
-                                  )
-                                : widget.sport == 6
-                                    ? Text(
-                                        '${context.watch<QuarterTabletennis>().getQuarter}',
-                                        style: const TextStyle(fontSize: 30),
-                                      )
-                                    : widget)
-      ],
+    return ElevatedButton(
+      style: ElevatedButton.styleFrom(
+        foregroundColor: Colors.white,
+        shape: const StadiumBorder(),
+        backgroundColor: const Color.fromARGB(255, 23, 36, 113),
+        minimumSize: const Size(90, 40),
+        // foreground
+      ),
+      onPressed: () {
+        if (widget.sport == 1) {
+          context.read<QuarterBasketball>().update();
+        }
+        if (widget.sport == 2) {
+          context.read<QuarterVolleyball>().update();
+        }
+        if (widget.sport == 3) {
+          context.read<QuarterSoccer>().update();
+        }
+        if (widget.sport == 4) {
+          context.read<QuarterFutsal>().update();
+        }
+        if (widget.sport == 5) {
+          context.read<QuarterBadminton>().update();
+        }
+        if (widget.sport == 6) {
+          context.read<QuarterTabletennis>().update();
+        }
+      },
+      child: const Text("Quarter"),
     );
   }
 }
